introduce ContextInferer to decouble ViewScopeRector from the context beeing inferred
